handling database conn, add sql backup file

+++ src/main/java/com/niko/servlets/DeleteCardServlet.java
@@ -15,19 +15,20 @@ import com.niko.dao.ApplicationDao;
  */
 public class DeleteCardServlet extends HttpServlet {
 
-
-	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+	protected void doGet(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
 
 	}
 
-
-	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+	protected void doPost(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
 		int cardId = Integer.parseInt(request.getParameter("cardId"));
 		ApplicationDao dao = new ApplicationDao();
 		Connection connection = (Connection) getServletContext().getAttribute("dbconnection");
-		
-		dao.deleteCard(cardId, connection);
-		
+
+		if (connection != null) {
+			dao.deleteCard(cardId, connection);
+		}
 		response.sendRedirect("/blog/HomeServlet");
 	}
