+++ src/main/java/com/niko/servlets/DeletePostServlet.java
@@ -10,26 +10,25 @@ import javax.servlet.http.HttpServletResponse;
 
 import com.niko.dao.ApplicationDao;
 
-
 public class DeletePostServlet extends HttpServlet {
 
+	protected void doPost(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
 
-
-	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		
 		int postId = Integer.parseInt(request.getParameter("postId"));
 		ApplicationDao dao = new ApplicationDao();
 		Connection connection = (Connection) getServletContext().getAttribute("dbconnection");
-		
-		dao.deletePost(postId, connection);
+
+		if (connection != null) {
+			dao.deletePost(postId, connection);
+		}
 
 		response.sendRedirect("/blog/BlogServlet");
-		
-		
+
 	}
 
-	
-	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		
+	protected void doGet(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
+
 	}
 }
